requirements.txt
asn1crypto==0.24.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
Django==2.2.2
django-environ==0.4.5
django-extensions==2.1.4
django-oauth-toolkit==1.2.0
djangorestframework==3.9.4
idna==2.8
oauthlib==3.1.0
Pillow==6.0.0
pycparser==2.19
PyJWT==1.7.1
pytz==2019.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
